+++ src/main/java/at/hannibal2/skyhanni/utils/BlockUtils.kt
@@ -1,6 +1,7 @@
 package at.hannibal2.skyhanni.utils
 
 import net.minecraft.block.Block
+import net.minecraft.block.state.IBlockState
 import net.minecraft.client.Minecraft
 import net.minecraft.tileentity.TileEntitySkull
 import net.minecraft.util.BlockPos
@@ -9,7 +10,10 @@ import net.minecraftforge.common.util.Constants
 object BlockUtils {
 
     fun LorenzVec.getBlockAt(): Block =
-        Minecraft.getMinecraft().theWorld.getBlockState(toBlocPos()).block
+        getBlockStateAt().block
+
+    fun LorenzVec.getBlockStateAt(): IBlockState =
+        Minecraft.getMinecraft().theWorld.getBlockState(toBlocPos())
 
     fun LorenzVec.isInLoadedChunk(): Boolean =
         Minecraft.getMinecraft().theWorld.chunkProvider.provideChunk(toBlocPos()).isLoaded